scale rack thumbs on server side, not in browser
There are two scales for the rack PNG thumbnail: 1:1 (used in rackspace
view) and N:1 (used in rack row view according to ROW_SCALE setting).
The latter for a long time remained the same as the former on the server
side (handled through "minirack" request subtype), relying on image
scaling to be performed by the browser. However, modern browsers tend to
scale images with resampling algorithms, not resizing, making the scaled
image look blurry.

This change adds an extra "midirack" request subtype standing for the
N:1 PNG image scaled on the server side, so that it looks sharp again.
The server-scaled images are not cached, which should not be a
performance issue since they are requested only for a specific rack row.

* printRackThumbImage(): add optional image scaling
* dispatchImageRequest(): add new "midirack" subtype
* renderRow(): pass the required image scale through the request